@@ -4,7 +4,7 @@ version: 1.0.5
 homepage: https://github.com/vanshg395/bubble_timeline
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
