@@ -2,16 +2,16 @@ import 'package:flutter/material.dart';
 
 class TimelineItem {
   final String title;
-  final String subtitle;
-  final String description;
+  final String? subtitle;
+  final String? description;
   final Widget child;
   final Color bubbleColor;
 
   const TimelineItem({
-    @required this.title,
+    /*required*/ required this.title,
     this.subtitle,
     this.description,
-    @required this.child,
-    @required this.bubbleColor,
+    /*required*/ required this.child,
+    /*required*/ required this.bubbleColor,
   });
 }
